% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-auc.R
\name{calculate_auc}
\alias{calculate_auc}
\title{Calculate different estimates related to AUC}
\usage{
calculate_auc(ya, yb)
}
\arguments{
\item{ya}{A vector responses at each level of covariates for treatment A.}

\item{yb}{A vector responses at each level of covariates for treatment B.}
}
\value{
A list of AUC estimates.
}
\description{
Calculate variance of predicted AUC, logit of predicted AUC, and variance of logit of predicted AUC responses passed
to calculate_auc variables to snake_case, even if it's already in snake_case.
}
\examples{
calculate_auc(ya = rnorm(100), yb = rnorm(100))
}
\author{
Som Bohora
}

