% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-auc.R
\name{compute_auc}
\alias{compute_auc}
\title{Compute AUC and its variance}
\usage{
compute_auc(d, nd)
}
\arguments{
\item{d}{A vector of response from treatment group}

\item{nd}{A vector of response from control group}
}
\description{
A function compute to compute AUC and its variance
}
\examples{
compute_auc(d = c(0.3,0.2,0.6), nd = c(0.2, 0.9,0.1, 0.5))
}
\author{
Som Bohora
}
