% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-inverse.R
\name{compute_inverse}
\alias{compute_inverse}
\title{Compute inverse of a function}
\usage{
compute_inverse(x)
}
\arguments{
\item{x}{A value between 0 and 1}
}
\description{
Ask for a function and compute its inverse
}
\examples{
compute_inverse(0.6)
}
\author{
Som Bohora
}
