% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fasd}
\alias{fasd}
\title{FASD data about effects of brochure types on alcohol use}
\format{A data frame with 210 rows and 4 variables:
\describe{
  \item{y}{alcohol use, number of drinks}
  \item{group}{types of brochures}
  \item{x1}{Whether vitamin was taken or not}
  \item{x2}{Smoking status}
}}
\source{
\url{http://www.netfas.net/}
}
\usage{
data(fasd)
}
\description{
A dataset containing the alcohol use and other attributes.
}
\examples{
data(fasd)
}
