% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-sAUC-shiny-app.R
\name{run_sAUC_shiny_app}
\alias{run_sAUC_shiny_app}
\title{Run Shiny application built for sAUC package (sAUC::run_sAUC_shiny_app())}
\usage{
run_sAUC_shiny_app()
}
\description{
This function allows users to run Shiny application built for sAUC package
where users can perform sAUC data analysis and run simulation for one predictor
}
\author{
Som Bohora
}
