% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.R
\name{sAUC}
\alias{sAUC}
\title{Fitting semiparametric AUC regression model adjusting for categorical covariates}
\usage{
sAUC(formula = FALSE, treatment_group = FALSE, data = FALSE)
}
\arguments{
\item{formula}{A formula object with response and covariates such as \code{response ~ x1 + x2}}

\item{treatment_group}{A treatment group for which a comparision is to be made}

\item{data}{A dataframe that contains variables needed for the analysis.}
}
\value{
A list of model summary, coefficients, AUC details, and session information.
}
\description{
\code{sAUC} is used to fit semiparametric AUC regression model specified by
giving a formula object of response and covariates and a separate argument of treatment
group. It will convert variables other than response into factors, estimate model parameters,
and display results.
}
\examples{
ds <- NULL
for (x1 in 0:1){
 for (x2 in 0:2){
   for (x3 in 0:2){
     for (group in 0:1){
         response <- round(rnorm(n = 100, mean = 0, sd = 1),4)
         column <- cbind(x1,x2,x3, group, response)
         ds <- as.data.frame(rbind(ds, column))
   }
  }
 }
}
ds[,c("x1", "x2", "x3", "group")] <- lapply(ds[,c("x1", "x2", "x3", "group")],
                                          function(x) factor(x))

sAUC(formula = response ~ x1 + x2 + x3, treatment_group = "group", data = ds)
}
\author{
Som Bohora
}
