% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-auc.R
\name{calculate_auc}
\alias{calculate_auc}
\title{Calculate different estimates related to AUC}
\usage{
calculate_auc(x, y = NULL, data = NULL)
}
\arguments{
\item{x}{A formula with two vector responses (For e.g., y1 ~ y2) or a vector of responses.}

\item{y}{A vector responses.}

\item{data}{An R dataframe object.}
}
\value{
A list of AUC estimates.
}
\description{
Calculate variance of predicted AUC, logit of predicted AUC, and variance of logit of predicted AUC responses passed
}
\examples{
calculate_auc(mpg ~ am, data = mtcars)
calculate_auc(iris$Sepal.Length, iris$Petal.Length)
}
\author{
Som Bohora
}

