% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-auc.R
\name{calculate_auc}
\alias{calculate_auc}
\title{This function calculates different estimates related to AUC}
\usage{
calculate_auc(x, y = NULL, data = NULL)
}
\arguments{
\item{x}{A formula object with two vector responses (For e.g., y1 ~ y2) or a vector of responses.}

\item{y}{A vector of responses.}

\item{data}{An R dataframe object.}
}
\value{
A list of AUC estimates.
}
\description{
This function calculates variance of predicted AUC, logit of predicted AUC,
and variance of logit of predicted AUC responses passed
}
\examples{
calculate_auc(mpg ~ am, data = datasets::mtcars)
calculate_auc(datasets::iris$Sepal.Length, datasets::iris$Petal.Length)
}
\author{
Som Bohora
}
