% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-auc.R, R/compute-inverse.R
\name{compute_inverse}
\alias{compute_inverse}
\alias{compute_auc}
\alias{compute_inverse}
\title{Compute AUC and its variance}
\usage{
compute_auc(d, nd)

compute_inverse(x)
}
\arguments{
\item{d}{A vector of response from treatment group}

\item{nd}{A vector of response from control group}

\item{x}{A value between 0 and 1}
}
\description{
A function compute to compute AUC and its variance

Ask for a function and compute its inverse
}
\examples{
compute_auc(d = c(0.3,0.2,0.6), nd = c(0.2, 0.9,0.1, 0.5))
compute_inverse(0.6)
}
\author{
Som Bohora

Som Bohora
}
