% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAUC.R
\name{sAUC}
\alias{sAUC}
\title{Run semiparametric AUC regression model adjusting for categorical covariates}
\usage{
sAUC(x = FALSE, treatment_group = FALSE, data = FALSE)
}
\arguments{
\item{x}{A formula with response and covariates such as response ~ x1 + x2}

\item{treatment_group}{A treatment group for which a comparision is to be made}

\item{data}{A dataframe that contains only variables needed for the analysis. At this point,
this dataframe should not contain any extra variables}
}
\value{
A list of model summary, coefficients, AUC details, and session information.
}
\description{
Ask for data frame that contains only required variables in the model,
Request to define response and treatment group, convert variables other than response
into factors, estimate model parameters, and display results.
}
\examples{
ds <- NULL
for (x1 in 0:1){
 for (x2 in 0:2){
   for (x3 in 0:2){
     for (group in 0:1){
         response <- round(rnorm(n = 100, mean = 0, sd = 1),4)
         column <- cbind(x1,x2,x3, group, response)
         ds <- as.data.frame(rbind(ds, column))
   }
  }
 }
}
ds[,c("x1", "x2", "x3", "group")] <- lapply(ds[,c("x1", "x2", "x3", "group")], function(x) factor(x))

sAUC(x = response ~ x1 + x2 + x3, treatment_group = "group", data = ds)

sAUC(x = y ~ x1 + x2, treatment_group = "group", data = fasd)
}
\author{
Som Bohora
}

