% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-one-predictor.R
\name{simulate_one_predictor}
\alias{simulate_one_predictor}
\title{Simulate Semi-parametric AUC regression adjusting for one categorical predictor}
\usage{
simulate_one_predictor(iter = 500, m = 100, p = 120)
}
\arguments{
\item{iter}{Number of realizations to run}

\item{m}{Number of observations on treatment condition}

\item{p}{number of observations on control condition}
}
\description{
Ask for number of iterations to run the simulation for Semiparametric AUC regression
adjusting for one discrete covariate. In this simulation, true parameters are as follows: beta0 = 0.15,
beta1 = 0.50, beta2 = 1.
}
\examples{
simulate_one_predictor(iter = 200, m = 100, p = 120)
}
\author{
Som Bohora
}
