% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-one-predictor.R
\name{simulate_one_predictor}
\alias{simulate_one_predictor}
\title{Simulate Semi-parametric AUC regression adjusting for one categorical predictor}
\usage{
simulate_one_predictor(iter = 100, m = 20, p = 30, b0 = 0.15,
  b1 = 0.5, b2 = 1)
}
\arguments{
\item{iter}{Number of realizations to be run}

\item{m}{Number of observations in treatment condition}

\item{p}{Number of observations in control condition}

\item{b0}{True intercept value}

\item{b1}{True beta 1 value}

\item{b2}{True beta 2 value}
}
\description{
It asks for number of iterations to be run, number of observations in treatment
and control groups for the simulation of Semiparametric AUC regression adjusting for one discrete
covariate. In this simulation, true model parameters are as follows: \eqn{\beta_0} = 0.15,
\eqn{\beta_1} = 0.50, \eqn{\beta_2} = 1.
}
\examples{
simulate_one_predictor(iter = 200, m = 100, p = 120)
}
\author{
Som Bohora
}
